- moves tests to the right location after typescript constants change
