+++ tests/Kiota.Builder.Tests/Writers/TypeScript/CodeConstantWriterTests.cs
@@ -3,8 +3,8 @@ using System.IO;
 using System.Linq;
 using Kiota.Builder.CodeDOM;
 using Kiota.Builder.Writers;
-using Kiota.Builder.Writers.Go;
 using Kiota.Builder.Writers.TypeScript;
+using Kiota.Builder.Extensions;
 
 using Xunit;
 
@@ -15,17 +15,35 @@ public sealed class CodeConstantWriterTests : IDisposable
     private const string DefaultName = "name";
     private readonly StringWriter tw;
     private readonly LanguageWriter writer;
-    private readonly CodeEnum currentEnum;
+    private CodeEnum currentEnum;
     private const string EnumName = "someEnum";
     private readonly CodeConstantWriter codeConstantWriter;
-
+    private readonly CodeNamespace root = CodeNamespace.InitRootNamespace();
+    private readonly CodeClass parentClass;
+    private const string MethodName = "methodName";
+    private const string ReturnTypeName = "Somecustomtype";
     public CodeConstantWriterTests()
     {
         writer = LanguageWriter.GetLanguageWriter(GenerationLanguage.TypeScript, DefaultPath, DefaultName);
         codeConstantWriter = new CodeConstantWriter(new());
         tw = new StringWriter();
         writer.SetTextWriter(tw);
-        var root = CodeNamespace.InitRootNamespace();
+        parentClass = root.AddClass(new CodeClass
+        {
+            Name = "parentClass",
+            Kind = CodeClassKind.RequestBuilder
+        }).First();
+        method = parentClass.AddMethod(new CodeMethod
+        {
+            Name = MethodName,
+            ReturnType = new CodeType
+            {
+                Name = ReturnTypeName
+            }
+        }).First();
+    }
+    private void AddCodeEnum()
+    {
         currentEnum = root.AddEnum(new CodeEnum
         {
             Name = EnumName,
@@ -65,6 +83,7 @@ public sealed class CodeConstantWriterTests : IDisposable
     [Fact]
     public void WritesEnumOptionDescription()
     {
+        AddCodeEnum();
         var option = new CodeEnumOption
         {
             Documentation = new()
@@ -83,6 +102,7 @@ public sealed class CodeConstantWriterTests : IDisposable
     [Fact]
     public void WritesEnum()
     {
+        AddCodeEnum();
         const string optionName = "option1";
         currentEnum.AddOption(new CodeEnumOption { Name = optionName });
         codeConstantWriter.WriteCodeElement(currentEnum.CodeEnumObject, writer);
@@ -97,8 +117,310 @@ public sealed class CodeConstantWriterTests : IDisposable
     [Fact]
     public void DoesntWriteAnythingOnNoOption()
     {
+        AddCodeEnum();
         codeConstantWriter.WriteCodeElement(currentEnum.CodeEnumObject, writer);
         var result = tw.ToString();
         Assert.Empty(result);
+    }
+    private readonly CodeMethod method;
+    [Fact]
+    public void DoesNotCreateDictionaryOnEmptyErrorMapping()
+    {
+        method.Kind = CodeMethodKind.RequestExecutor;
+        method.HttpMethod = HttpMethod.Get;
+
+        AddRequestBodyParameters();
+        var constant = CodeConstant.FromRequestBuilderToRequestsMetadata(parentClass);
+        parentClass.GetImmediateParentOfType<CodeNamespace>().AddConstant(constant);
+        writer.Write(constant);
+        var result = tw.ToString();
+        Assert.DoesNotContain("errorMappings", result);
+        AssertExtensions.CurlyBracesAreClosed(result);
+    }
+    [Fact]
+    public void WritesRequestGeneratorBodyForMultipart()
+    {
+        method.Kind = CodeMethodKind.RequestGenerator;
+        method.HttpMethod = HttpMethod.Post;
+        AddRequestProperties();
+        AddRequestBodyParameters();
+        method.Parameters.First(static x => x.IsOfKind(CodeParameterKind.RequestBody)).Type = new CodeType { Name = "MultipartBody", IsExternal = true };
+        var constant = CodeConstant.FromRequestBuilderToRequestsMetadata(parentClass);
+        parentClass.GetImmediateParentOfType<CodeNamespace>().AddConstant(constant);
+        writer.Write(constant);
+        var result = tw.ToString();
+        Assert.Contains("setContentFromParsable", result);
+        AssertExtensions.CurlyBracesAreClosed(result);
+    }
+    [Fact]
+    public void WritesRequestExecutorBodyForCollections()
+    {
+        method.Kind = CodeMethodKind.RequestExecutor;
+        method.HttpMethod = HttpMethod.Get;
+        method.ReturnType.CollectionKind = CodeTypeBase.CodeTypeCollectionKind.Array;
+        AddRequestBodyParameters();
+        var constant = CodeConstant.FromRequestBuilderToRequestsMetadata(parentClass);
+        parentClass.GetImmediateParentOfType<CodeNamespace>().AddConstant(constant);
+        writer.Write(constant);
+        var result = tw.ToString();
+        Assert.Contains("sendCollectionAsync", result);
+        AssertExtensions.CurlyBracesAreClosed(result);
+    }
+    [Fact]
+    public void WritesRequestGeneratorBodyForScalar()
+    {
+        method.Kind = CodeMethodKind.RequestGenerator;
+        method.HttpMethod = HttpMethod.Get;
+        AddRequestProperties();
+        AddRequestBodyParameters();
+        method.AcceptedResponseTypes.Add("application/json");
+        var constant = CodeConstant.FromRequestBuilderToRequestsMetadata(parentClass);
+        parentClass.GetImmediateParentOfType<CodeNamespace>().AddConstant(constant);
+        writer.Write(constant);
+        var result = tw.ToString();
+        Assert.Contains("const requestInfo = new RequestInformation(HttpMethod.GET, this.urlTemplate, this.pathParameters)", result);
+        Assert.Contains("requestInfo.headers.tryAdd(\"Accept\", \"application/json\")", result);
+        Assert.Contains("requestInfo.configure", result);
+        Assert.Contains("setContentFromScalar", result);
+        Assert.Contains("return requestInfo;", result);
+        AssertExtensions.CurlyBracesAreClosed(result);
+    }
+    [Fact]
+    public void WritesRequestGeneratorBodyForParsable()
+    {
+        method.Kind = CodeMethodKind.RequestGenerator;
+        method.HttpMethod = HttpMethod.Get;
+        AddRequestProperties();
+        AddRequestBodyParameters(true);
+        method.AcceptedResponseTypes.Add("application/json");
+        var constant = CodeConstant.FromRequestBuilderToRequestsMetadata(parentClass);
+        parentClass.GetImmediateParentOfType<CodeNamespace>().AddConstant(constant);
+        writer.Write(constant);
+        var result = tw.ToString();
+        Assert.Contains("const requestInfo = new RequestInformation(HttpMethod.GET, this.urlTemplate, this.pathParameters", result);
+        Assert.Contains("requestInfo.headers.tryAdd(\"Accept\", \"application/json\")", result);
+        Assert.Contains("requestInfo.configure", result);
+        Assert.Contains("setContentFromParsable", result);
+        Assert.Contains("return requestInfo;", result);
+        AssertExtensions.CurlyBracesAreClosed(result);
+    }
+    [Fact]
+    public void WritesRequestGeneratorBodyKnownRequestBodyType()
+    {
+        method.Kind = CodeMethodKind.RequestGenerator;
+        method.HttpMethod = HttpMethod.Get;
+        AddRequestProperties();
+        AddRequestBodyParameters(true);
+        method.Kind = CodeMethodKind.RequestGenerator;
+        method.HttpMethod = HttpMethod.Post;
+        AddRequestProperties();
+        AddRequestBodyParameters(false);
+        method.Parameters.OfKind(CodeParameterKind.RequestBody).Type = new CodeType
+        {
+            Name = new TypeScriptConventionService().StreamTypeName,
+            IsExternal = true,
+        };
+        method.RequestBodyContentType = "application/json";
+        var constant = CodeConstant.FromRequestBuilderToRequestsMetadata(parentClass);
+        parentClass.GetImmediateParentOfType<CodeNamespace>().AddConstant(constant);
+        writer.Write(constant);
+        var result = tw.ToString();
+        Assert.Contains("setStreamContent", result, StringComparison.OrdinalIgnoreCase);
+        Assert.Contains("application/json", result, StringComparison.OrdinalIgnoreCase);
+        AssertExtensions.CurlyBracesAreClosed(result);
+    }
+    [Fact]
+    public void WritesRequestGeneratorBodyUnknownRequestBodyType()
+    {
+        method.Kind = CodeMethodKind.RequestGenerator;
+        method.HttpMethod = HttpMethod.Get;
+        AddRequestProperties();
+        AddRequestBodyParameters(true);
+        method.Kind = CodeMethodKind.RequestGenerator;
+        method.HttpMethod = HttpMethod.Post;
+        AddRequestProperties();
+        AddRequestBodyParameters(false);
+        method.Parameters.OfKind(CodeParameterKind.RequestBody).Type = new CodeType
+        {
+            Name = new TypeScriptConventionService().StreamTypeName,
+            IsExternal = true,
+        };
+        method.AddParameter(new CodeParameter
+        {
+            Name = "requestContentType",
+            Type = new CodeType()
+            {
+                Name = "string",
+                IsExternal = true,
+            },
+            Kind = CodeParameterKind.RequestBodyContentType,
+        });
+        var constant = CodeConstant.FromRequestBuilderToRequestsMetadata(parentClass);
+        parentClass.GetImmediateParentOfType<CodeNamespace>().AddConstant(constant);
+        writer.Write(constant);
+        var result = tw.ToString();
+        Assert.Contains("setStreamContent", result, StringComparison.OrdinalIgnoreCase);
+        Assert.DoesNotContain("application/json", result, StringComparison.OrdinalIgnoreCase);
+        Assert.Contains(", requestContentType", result, StringComparison.OrdinalIgnoreCase);
+        AssertExtensions.CurlyBracesAreClosed(result);
+    }
+    [Fact]
+    public void WritesRequestExecutorBody()
+    {
+        method.Kind = CodeMethodKind.RequestExecutor;
+        method.HttpMethod = HttpMethod.Get;
+        var error4XX = root.AddClass(new CodeClass
+        {
+            Name = "Error4XX",
+        }).First();
+        var error5XX = root.AddClass(new CodeClass
+        {
+            Name = "Error5XX",
+        }).First();
+        var error403 = root.AddClass(new CodeClass
+        {
+            Name = "Error403",
+        }).First();
+        method.AddErrorMapping("4XX", new CodeType { Name = "Error4XX", TypeDefinition = error4XX });
+        method.AddErrorMapping("5XX", new CodeType { Name = "Error5XX", TypeDefinition = error5XX });
+        method.AddErrorMapping("403", new CodeType { Name = "Error403", TypeDefinition = error403 });
+        AddRequestBodyParameters();
+        Assert.Throws<InvalidOperationException>(() => writer.Write(method));
+    }
+    [Fact]
+    public void WritesIndexer()
+    {
+        AddRequestProperties();
+        method.Kind = CodeMethodKind.IndexerBackwardCompatibility;
+        method.OriginalIndexer = new()
+        {
+            Name = "indx",
+            ReturnType = new CodeType
+            {
+                Name = "string",
+            },
+            IndexParameter = new()
+            {
+                Name = "id",
+                SerializationName = "id",
+                Type = new CodeType
+                {
+                    Name = "string",
+                    IsNullable = true,
+                },
+            }
+        };
+        var constant = CodeConstant.FromRequestBuilderToRequestsMetadata(parentClass);
+        parentClass.GetImmediateParentOfType<CodeNamespace>().AddConstant(constant);
+        writer.Write(constant);
+        var result = tw.ToString();
+        Assert.Contains("this.requestAdapter", result);
+        Assert.Contains("this.pathParameters", result);
+        Assert.Contains("id", result);
+        Assert.Contains("return new", result);
+    }
+    [Fact]
+    public void WritesPathParameterRequestBuilder()
+    {
+        AddRequestProperties();
+        method.Kind = CodeMethodKind.RequestBuilderWithParameters;
+        method.AddParameter(new CodeParameter
+        {
+            Name = "pathParam",
+            Kind = CodeParameterKind.Path,
+            Type = new CodeType
+            {
+                Name = "string"
+            }
+        });
+        var constant = CodeConstant.FromRequestBuilderToRequestsMetadata(parentClass);
+        parentClass.GetImmediateParentOfType<CodeNamespace>().AddConstant(constant);
+        writer.Write(constant);
+        var result = tw.ToString();
+        Assert.Contains("this.requestAdapter", result);
+        Assert.Contains("this.pathParameters", result);
+        Assert.Contains("pathParam", result);
+        Assert.Contains("return new", result);
+    }
+    private void AddRequestProperties()
+    {
+        parentClass.AddProperty(new CodeProperty
+        {
+            Name = "requestAdapter",
+            Kind = CodePropertyKind.RequestAdapter,
+            Type = new CodeType
+            {
+                Name = "RequestAdapter",
+            },
+        });
+        parentClass.AddProperty(new CodeProperty
+        {
+            Name = "pathParameters",
+            Kind = CodePropertyKind.PathParameters,
+            Type = new CodeType
+            {
+                Name = "string"
+            },
+        });
+        parentClass.AddProperty(new CodeProperty
+        {
+            Name = "urlTemplate",
+            Kind = CodePropertyKind.UrlTemplate,
+            Type = new CodeType
+            {
+                Name = "string"
+            },
+        });
+    }
+    private void AddRequestBodyParameters(bool useComplexTypeForBody = false)
+    {
+        var stringType = new CodeType
+        {
+            Name = "string",
+        };
+        var requestConfigClass = parentClass.AddInnerClass(new CodeClass
+        {
+            Name = "RequestConfig",
+            Kind = CodeClassKind.RequestConfiguration,
+        }).First();
+        requestConfigClass.AddProperty(new()
+        {
+            Name = "h",
+            Kind = CodePropertyKind.Headers,
+            Type = stringType,
+        },
+        new()
+        {
+            Name = "q",
+            Kind = CodePropertyKind.QueryParameters,
+            Type = stringType,
+        },
+        new()
+        {
+            Name = "o",
+            Kind = CodePropertyKind.Options,
+            Type = stringType,
+        });
+        method.AddParameter(new CodeParameter
+        {
+            Name = "b",
+            Kind = CodeParameterKind.RequestBody,
+            Type = useComplexTypeForBody ? new CodeType
+            {
+                Name = "SomeComplexTypeForRequestBody",
+                TypeDefinition = TestHelper.CreateModelClass(root, "SomeComplexTypeForRequestBody"),
+            } : stringType,
+        });
+        method.AddParameter(new CodeParameter
+        {
+            Name = "c",
+            Kind = CodeParameterKind.RequestConfiguration,
+            Type = new CodeType
+            {
+                Name = "RequestConfig",
+                TypeDefinition = requestConfigClass,
+            },
+            Optional = true,
+        });
     }
 }
